deleteボタン機能ロジック、modal親から渡すよう修正
