+++ frontend/components/admin/pages/AdvertisementEditor.tsx
@@ -1,5 +1,6 @@
 'use client'
 import React, { useState, FormEvent } from 'react'
+import { useRouter, useParams } from 'next/navigation'
 
 import {
   Box,
@@ -26,6 +27,8 @@ import { THEME_COLOR } from '../../../constants'
 import { AdvertisementType } from '../../../types'
 import { advertisementSchema } from '../../../zod'
 import { ZodIssue } from 'zod'
+import { useCustomToast } from '../../../hooks/useCustomToast'
+import { PATHS } from '../../../constants'
 
 type AdvertisementEditorProps = {
   advertisement: AdvertisementType
@@ -124,6 +127,44 @@ export function AdvertisementEditor({
       })
     } finally {
       SetIsSubmitting(false)
+    }
+  }
+  const { showSuccessToast, showErrorToast } = useCustomToast()
+
+  const { isOpen, onOpen, onClose } = useDisclosure()
+  const [showModalContent, setShowModalContent] = useState(true)
+
+  const router = useRouter()
+  const params = useParams()
+  const advertisementId = params.advertisementId
+
+  const deleteAdvertisement = async () => {
+    try {
+      const response = await fetch(
+        `${process.env.NEXT_PUBLIC_BACKEND_URL}/advertisement`,
+        {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            advertisementId,
+          }),
+        },
+      )
+
+      const result = await response.json()
+      if (response.ok) {
+        showSuccessToast(result.message)
+        setShowModalContent(false)
+        setTimeout(() => {
+          router.push(PATHS.ADMIN.ADVERTISEMENT.LIST.path)
+        }, 4000)
+      } else {
+        showErrorToast(result.message)
+      }
+    } catch (error) {
+      showErrorToast('広告の削除に失敗しました。')
     }
   }
   return (
